Move Railtie to own file
